+++ projects/myrmidon/cadmus-part-general-ui/src/lib/pin-links-fragment/pin-links-fragment.component.ts
@@ -122,6 +122,15 @@ export class PinLinksFragmentComponent
       validators: NgxToolsValidators.strictMinLengthValidator(1),
       nonNullable: true,
     });
+    // settings
+    this.initSettings<PinLinksFragmentSettings>(
+      PIN_LINKS_FRAGMENT_TYPEID,
+      (settings) => {
+        this.lookupProviderOptions.set(
+          settings?.lookupProviderOptions || undefined,
+        );
+      },
+    );
   }
 
   public override ngOnInit(): void {
@@ -187,16 +196,6 @@ export class PinLinksFragmentComponent
     if (data?.thesauri) {
       this.updateThesauri(data.thesauri);
     }
-    // settings
-    this._appRepository
-      ?.getSettingFor<PinLinksFragmentSettings>(
-        PIN_LINKS_FRAGMENT_TYPEID,
-        this.identity()?.roleId || undefined,
-      )
-      .then((settings) => {
-        const options = settings?.lookupProviderOptions;
-        this.lookupProviderOptions.set(options || undefined);
-      });
     // form
     this.updateForm(data?.value);
   }